backend/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
python-dotenv==1.2.1
google-generativeai==0.8.5
pymupdf==1.26.7
requests==2.32.5
msal==1.34.0
keyring==25.7.0
pydantic==2.12.4
python-multipart
slowapi==0.1.9
defusedxml==0.7.1
Pillow==11.1.0
Office365-REST-Python-Client==2.6.2
flashtext==2.7
keyrings.alt==5.0.0
sqlalchemy==2.0.25
PyJWT==2.8.0
cryptography==42.0.5
